Add Cypress test suite and update related configurations
Six new files were added, primarily related to the integration of the Cypress testing suite. This includes support/commands.ts and support/e2e.ts which store command declarations and end-to-end testing configurations respectively. Additionally, an Angular dev server was defined in cypress.config.ts, and an example test was written under the Cypress end-to-end directory. Changes were also made to the file tsconfig.json and the package-lock.json file was updated to include references to the new files and packages.